Consolidate redundant 'server' packages
