+++ src/test/java/com/eighthlight/fabrial/test/server/MockCloseable.java
@@ -1,4 +1,4 @@
-package com.eighthlight.fabrial.test.http.server;
+package com.eighthlight.fabrial.test.server;
 
 import com.eighthlight.fabrial.server.CloseStateSupplier;
