+++ src/test/java/com/eighthlight/fabrial/test/server/MockSocketController.java
@@ -1,4 +1,4 @@
-package com.eighthlight.fabrial.test.http.server;
+package com.eighthlight.fabrial.test.server;
 
 import com.eighthlight.fabrial.server.ClientConnection;
 import com.eighthlight.fabrial.server.SocketController;
